Move Ui to ui artifact
