+++ circuit-codegen-annotations/src/commonMain/kotlin/com/slack/circuit/codegen/annotations/CircuitInject.kt
@@ -5,7 +5,7 @@ package com.slack.circuit.codegen.annotations
 import androidx.compose.runtime.Composable
 import com.slack.circuit.CircuitConfig
 import com.slack.circuit.Presenter
-import com.slack.circuit.Ui
+import com.slack.circuit.runtime.ui.Ui
 import com.slack.circuit.runtime.CircuitUiState
 import com.slack.circuit.runtime.Navigator
 import com.slack.circuit.runtime.Screen
