+++ circuit/src/commonMain/kotlin/com/slack/circuit/EventListener.kt
@@ -6,6 +6,7 @@ import com.slack.circuit.runtime.CircuitContext
 import com.slack.circuit.runtime.CircuitUiState
 import com.slack.circuit.runtime.Navigator
 import com.slack.circuit.runtime.Screen
+import com.slack.circuit.runtime.ui.Ui
 
 /**
  * A listener for tracking the state changes of a given [Screen]. This can be used for
