+++ circuit/src/jvmTest/kotlin/com/slack/circuit/EventListenerTest.kt
@@ -14,6 +14,7 @@ import com.slack.circuit.runtime.CircuitContext
 import com.slack.circuit.runtime.CircuitUiState
 import com.slack.circuit.runtime.Navigator
 import com.slack.circuit.runtime.Screen
+import com.slack.circuit.runtime.ui.Ui
 import java.time.LocalTime
 import java.time.ZoneOffset.UTC
 import java.util.concurrent.TimeUnit
