+++ samples/counter/desktop/src/main/kotlin/com/slack/circuit/sample/counter/desktop/DesktopCounterCircuit.kt
@@ -25,14 +25,14 @@ import androidx.compose.ui.window.singleWindowApplication
 import com.slack.circuit.CircuitCompositionLocals
 import com.slack.circuit.CircuitConfig
 import com.slack.circuit.CircuitContent
-import com.slack.circuit.Ui
+import com.slack.circuit.runtime.ui.Ui
 import com.slack.circuit.runtime.CircuitContext
 import com.slack.circuit.runtime.Screen
 import com.slack.circuit.sample.counter.CounterEvent
 import com.slack.circuit.sample.counter.CounterPresenterFactory
 import com.slack.circuit.sample.counter.CounterScreen
 import com.slack.circuit.sample.counter.CounterState
-import com.slack.circuit.ui
+import com.slack.circuit.runtime.ui.ui
 
 object DesktopCounterScreen : CounterScreen
