+++ samples/star/src/main/kotlin/com/slack/circuit/star/di/CircuitModule.kt
@@ -4,7 +4,7 @@ package com.slack.circuit.star.di
 
 import com.slack.circuit.CircuitConfig
 import com.slack.circuit.Presenter
-import com.slack.circuit.Ui
+import com.slack.circuit.runtime.ui.Ui
 import com.squareup.anvil.annotations.ContributesTo
 import dagger.Module
 import dagger.Provides
